// NARANJA/designsystem-core/src/lib/src/utils/enums/colors.enum.ts
export enum ZumoColors {
  Brand_200 = '#FEE4D9',
  Brand_300 = '#FEC1A6',
  Brand_400 = '#FE9E73',
  Brand_500 = '#FE5000',
  Brand_600 = '#CB4000',
  Brand_700 = '#983000',
  Brand_800 = '#7E2700',

  Grayscale_100 = '#FFFFFF',
  Grayscale_200 = '#F2F2F2',
  Grayscale_300 = '#DDDDDD',
  Grayscale_400 = '#BBBBBB',
  Grayscale_500 = '#949494',
  Grayscale_600 = '#767676',
  Grayscale_700 = '#666666',
  Grayscale_800 = '#333333',
  Grayscale_900 = '#000000',

  Warning_200 = '#FFF3DB',
  Warning_300 = '#FFE4AB',
  Warning_400 = '#FFD987',
  Warning_500 = '#FFB30F',
  Warning_600 = '#996B09',
  Warning_700 = '#735006',
  Warning_800 = '#5A3F05',

  Error_200 = '#FCE5E5',
  Error_300 = '#F6B3B3',
  Error_400 = '#F18080',
  Error_500 = '#E40000',
  Error_600 = '#C20000',
  Error_700 = '#A00000',
  Error_800 = '#880000',

  Success_200 = '#E6F8EA',
  Success_300 = '#AAE8B8',
  Success_400 = '#7ADB8F',
  Success_500 = '#0EBF34',
  Success_600 = '#098524',
  Success_700 = '#075F1A',
  Success_800 = '#054C14',

  Info_200 = '#E6EFFB',
  Info_300 = '#BFD6F7',
  Info_400 = '#72A5ED',
  Info_500 = '#005DE0',
  Info_600 = '#004AB3',
  Info_700 = '#003786',
  Info_800 = '#002E70'
}